Extract code into its own file
